back_end/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
astroid==3.3.11
attrs==25.4.0
certifi==2025.11.12
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fastapi==0.116.1
fire==0.7.1
fonttools==4.60.0
h11==0.16.0
idna==3.10
ijson==2.3
img2pdf==0.6.1
inflect==7.5.0
isort==6.0.1
lxml==5.3.2
mando==0.7.1
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
numpy==2.2.6
ocrmypdf==16.11.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pdf2docx==0.5.8
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.1.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.8
PyMuPDF==1.26.4
pyOpenSSL==25.1.0
pyravendb==5.0.0.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2025.2
radon==6.0.1
requests==2.32.4
requests-futures==1.0.2
requests-pkcs12==1.25
rich==14.1.0
selenium==4.38.0
sherlock-project==0.16.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.47.3
stem==1.8.2
termcolor==3.1.0
tomlkit==0.13.3
trio==0.32.0
trio-websocket==0.12.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
wrapt==1.17.3
wsproto==1.3.1
